Moving hack.hands code to a separated module. Completing the connection view function.
